requirements.txt
alembic==0.9.3.dev0
amqp==2.2.2
anyjson==0.3.3
appdirs==1.4.3
apturl==0.5.2
asgiref==3.2.3
asn1crypto==0.24.0
astroid==2.3.3
autopep8==1.5
Babel==2.4.0
billiard==3.5.0.3
bleach==2.1.2
blinker==1.4
Brlapi==0.6.6
celery==4.1.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==6.7
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.1.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
Django==3.0.3
<<<<<<< HEAD
django-crispy-forms==1.9.0
django-widget-tweaks==1.4.6
=======
<<<<<<< HEAD
=======
<<<<<<< HEAD
>>>>>>> 3b5f5fa2c79af2e71417e3840390e733337e0fba
>>>>>>> 2ba867ca9b157e700814aa9a2461d554c10d5911
django-allauth==0.41.0
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-google-maps==0.12.1
django-rest-knox==4.1.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
empy==3.3.2
entrypoints==0.2.3.post1
ephem==3.7.6.0
evdev==0.7.0
filelock==3.0.12
Flask==0.12.2
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.4.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
gestures==0.2.2
html5lib==0.999999999
httplib2==0.9.2
<<<<<<< HEAD
spy-forms==1.9.0
=======
<<<<<<< HEAD
django-crispy-forms==1.9.0
django-widget-tweaks==1.4.6
=======
=======
django-crispy-forms==1.9.0
>>>>>>> 2ba867ca9b157e700814aa9a2461d554c10d5911
django-widget-tweaks==1.4.6
idna==2.9
importlib-metadata==1.5.0
importlib-resources==1.3.1
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.3.21
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
kombu==4.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
netifaces==0.10.4
notebook==5.2.2
numpy==1.18.1
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
pandas==0.25.3
pandocfilters==1.4.2
paramiko==2.0.0
passlib==1.7.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.1.0
pipenv==2018.11.26
prompt-toolkit==1.0.15
protobuf==3.0.0
psutil==5.4.2
psycopg2==2.8.4
pyasn1==0.4.2
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
pyinotify==0.9.6
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==2.4.6
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
python-memcached==1.53
python3-openid==3.1.0
pytz==2019.3
pyudev==0.21.0
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
reportlab==3.4.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
scipy==1.4.1
scour==0.36
seaborn==0.9.0
SecretStorage==2.3.1
simplegeneric==0.8.1
simplejson==3.13.2
six==1.14.0
SQLAlchemy==1.1.11
sqlparse==0.3.0
sshtunnel==0.1.4
system-service==0.3
systemd-python==234
terminado==0.7
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.3
vine==1.1.4
virtualenv==20.0.10
virtualenv-clone==0.5.3
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.1
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
